requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
spacy>=3.6.0
textblob>=0.17.0
lime>=0.2.0
shap>=0.42.0
streamlit-drawable-canvas>=0.9.0
Pillow>=10.0.0
plotly>=5.18.0
opencv-python-headless>=4.8.0
